Fixed glitch with ordering tracks

+++ app/controllers/setlists_controller.rb
@@ -9,46 +9,34 @@ class SetlistsController < ApplicationController
         render json: setlist, serializer: SetlistWithTracksSerializer
     end
 
-    # def update
-    #     if check_session
-    #         user = current_user
-    #         setlist = find_setlist
-
-    #         if user.id == setlist.user_id
-    #             setlist.update!(setlist_info_params)
-    #             render json: setlist, serializer: SetlistWithTracksSerializer
-    #         else
-    #             render_not_authorized_response
-    #         end
-    #     else
-    #         render_not_authorized_response
-    #     end
-    # end
-
     def update
         if check_session
             user = current_user
             @setlist = find_setlist
 
-
             if @setlist.user_id === user.id
                 ActiveRecord::Base.transaction do
-                    tracks = track_params[:tracks]
-                    tracks2 = tracks.map{|t| Track.where(name: t[:name], artist: t[:artist], genre: t[:genre], bpm: t[:bpm], length: t[:length]).first_or_create(t)}
-
-                    h = 1
-
-                    @updated = []
-                    tracks2.each do |t|
-                        m = @setlist.setlist_tracks.where(track_order: h).first_or_initialize
-                        m.track_id = t.id
-                        m.save
-                        h+=1
-                        @updated << m
+                    if check_update_params
+                        @setlist.update!(setlist_info_params)
+                    else
+                        tracks = track_params[:tracks]
+                        tracks2 = tracks.map{|t| where_duplicate(t).first_or_create(t)}
+    
+                        h = 1
+                        @updated = []
+    
+                        tracks2.each do |t|
+                            m = @setlist.setlist_tracks.where(track_order: h).first_or_initialize
+                            m.track_id = t.id
+                            m.save
+                            h+=1
+                            @updated << m
+                        end
+    
+                        ids = @updated.map{|i| i['id'].to_i}
+                        setlist_delete = @setlist.setlist_tracks.where.not(id: ids).destroy_all
+                        @setlist.update!(avg_bpm: @setlist.tracks.average(:bpm), length: @setlist.setlist_tracks.map {|t| t.track}.sum(&:length)/60)
                     end
-
-                    ids = @updated.map{|i| i['id'].to_i}
-                    setlist_delete = @setlist.setlist_tracks.where.not(id: ids).destroy_all
                 end
                 render json: @setlist, status: :accepted, serializer: SetlistWithTracksSerializer
             else
@@ -107,6 +95,10 @@ class SetlistsController < ApplicationController
 
     def where_duplicate(t)
         Track.where(name: t[:name], artist: t[:artist], genre: t[:genre], bpm: t[:bpm], length: t[:length])
+    end
+
+    def check_update_params
+        params.key?("name" && "description")
     end
 
     def setlist_params
